move registration methods to AttributeRegistration
